feat(v1.2.5): 全链路评估修复 — Web 控制台重建 + API 测试 58/58 覆盖 + 全量回归全绿
基于全量全链路开发进度评估的系统修复（含 v1.2.4 未提交存量）：

【Web 控制台恢复与重建（P0）】
- 从 git 历史恢复并重建被 0613533 删除的 web/ 控制台：19+ HTML 页面
  （workbench/login/admin/studio/demo 等）+ CSS/JS 资产 + 品牌资源 +
  法律页面 + A2UI 前端资源 + Flutter PWA 构建产物
- 新增 web/404.html 自定义错误页，nginx 双 server 配置 error_page 404

【API 测试覆盖率 21/58 → 58/58（P0）】
- 新增 44 个测试文件约 320 个测试函数，覆盖全部 API 模块
  CRUD / Auth / 越权 / 边界 case
- 修复 28 个因 API 演进过期的测试用例（灯具类型 CHECK 约束、
  健康监测 schema、变更单状态机、家具 PATCH 方法、Matter 501 等）
- 全量回归：1387 passed / 0 failed / 17 skipped

【功能补全（v1.2.4 存量）】
- sketch_to_3d 接入真实多模态视觉模型（deepseek→glm→qwen 优先级，
  sketch_to_3d_vision_enabled flag）
- Flutter 业务实体类型化模型 10 个（不可变模式 fromJson/toJson/copyWith）
- A2A 任务持久化：内存 dict → A2ATask 表（含 alembic 迁移 + TTL 24h）
- voice /process-enhanced LLM 语义意图路由（10 类意图，
  voice_llm_routing_enabled flag，失败降级关键词匹配）
- 回复模板系统 app/services/reply_templates.py 集中管理硬编码中文回复

【本轮代码修复】
- A2A /api/a2a/agents 列表序列化修复：Agent 类对象 → 可序列化字典结构
- ChangeOrderStateError 注册专属异常处理器：非法状态流转返回 409
  Conflict（而非 500）
- 版本号统一 1.2.5：.env.example / .env.production.example /
  deploy-production.sh / CI（3 处）

【项目清理】
- 删除 serverless/ FC 微服务死配置（模块化单体从未接条件路由加载）
- 删除冗余脚本：e2e-demo.sh / e2e-verify.py / test_agents.py /
  verify-ac.sh / e2e_ar_web.py

+++ app/schemas/matter_device.py
@@ -115,7 +115,7 @@ class MatterPlacementPlanResponse(BaseModel):
     protocol: str
     rooms: list[dict[str, Any]] = Field(default_factory=list)
     total_device_count: int = 0
-    estimated_power_w: int = 0
+    estimated_power_w: float = 0
     commissioning_guide: str = ""
     commissioned_devices: list[dict[str, Any]] = Field(
         default_factory=list, description="已配对的 Matter 设备列表"
